add true and trainable dynamics
