+++ phd_experiments/torch_ode/torch_ode_solver.py
@@ -8,7 +8,6 @@ import torch
 class TorchODESolverSolution:
     def __init__(self, zf: torch.Tensor, z_trajectory: Iterable[torch.Tensor], t_values: Iterable[float]):
         """
-
         Parameters
         ----------
         zf
@@ -24,7 +23,6 @@ class TorchODESolver(ABC):
         self.step_size = step_size
         self.logger = logging.getLogger()
         self.nfe = 0 # Number of Function Evaluations
-
 
     @abstractmethod
     def solve_ivp(self, func: Callable[[float, torch.Tensor, ...], torch.Tensor], t_span: Tuple,
