+++ phd_experiments/torch_ode/torch_ode_train.py
@@ -38,6 +38,8 @@ from torch.utils.data import TensorDataset, DataLoader
 from phd_experiments.torch_ode.torch_ode_solver import TorchODESolver
 from phd_experiments.torch_ode.torch_ode_utils import get_device_info, format_timedelta, log_train_experiment
 from phd_experiments.torch_ode.torch_rk45 import TorchRK45
+from phd_experiments.torch_ode.torch_train_learnable_dynamics import ODEFuncNN3Layer
+from phd_experiments.torch_ode.torch_train_true_dynamics import f_simple_linear_uncoupled_ode_1
 
 #########################
 # Global variables and settings
@@ -52,12 +54,6 @@ np.random.seed(SEED)
 
 
 # Data Generation methods and classes
-def f_simple_linear_uncoupled_ode_1(t: float, y: np.ndarray, a: float):
-    # http://web.math.ucsb.edu/~ebrahim/lin_ode_sys.pdf eqn (1)
-    yprime = np.empty(2)
-    yprime[0] = a * y[0]
-    yprime[1] = -y[1]
-    return yprime
 
 
 class ToyODEDataGenerator():
@@ -143,42 +139,6 @@ class TorchODETrainer:
         torch.save(obj=self.ode_func_model.state_dict(), f=model_path)
 
 
-class ODEFunc(nn.Module):
-    # copy from
-    def __init__(self, device: torch.device, tensor_dtype: torch.dtype):
-        super(ODEFunc, self).__init__()
-        self.nfe = 0  # Number of function evaluations
-        # https://github.com/rtqichen/torchdiffeq/blob/master/examples/odenet_mnist.py#L102
-        self.device = device
-        self.tensor_dtype = tensor_dtype
-        self.net = nn.Sequential(
-            nn.Linear(2, 50, device=self.device, dtype=self.tensor_dtype),
-            nn.Tanh(),
-            nn.Linear(50, 2, device=self.device, dtype=tensor_dtype),
-        )
-        self.net.cuda()  # move to cuda , might be redundant as device is set in nn.Linear
-        assert next(self.net.parameters()).is_cuda, "Model is not on Cuda"
-        for m in self.net.modules():
-            if isinstance(m, nn.Linear):
-                nn.init.normal_(m.weight, mean=0, std=0.1)
-                nn.init.constant_(m.bias, val=0)
-
-    def forward(self, t, y):
-        # https://github.com/rtqichen/torchdiffeq/blob/master/examples/odenet_mnist.py#L105
-        self.nfe += 1  # Update number of function evaluations (implicit depth)
-        y_pred = self.net(y)  # can be played with !
-        return y_pred
-
-    def get_nfe(self):
-        return self.nfe
-
-    def reset_nfe(self):
-        self.nfe = 0
-
-
-# Learn dynamics
-
-
 def evaluate(solver: TorchODESolver, ode_func_model: nn.Module, t_span: Tuple, test_set: DataLoader,
              test_loss_fn: Callable):
     batches_losses = []
@@ -221,10 +181,9 @@ if __name__ == '__main__':
     dataset_config = {'t_span': (0, 1), 'f_params': {'a': 1}, 'num_batches': 1000, 'batch_size': 32,
                       'splits': [0.7, 0.2, 0.1], 'f_true_dynamics': f_simple_linear_uncoupled_ode_1}
 
-    data_gen = ToyODEDataGenerator(device=torch_configs['device'], tensor_dtype=torch_configs['TENSOR_DTYPE'], ulow=-10,
-                                   uhigh=10,
-                                   f=f_simple_linear_uncoupled_ode_1, t_span=dataset_config['t_span'],
-                                   args=(dataset_config['f_params']['a'],))
+    data_gen = ToyODEDataGenerator(device=torch_configs['device'], tensor_dtype=torch_configs['TENSOR_DTYPE'],
+                                   ulow=-10, uhigh=10, f=dataset_config['f_true_dynamics'],
+                                   t_span=dataset_config['t_span'], args=(dataset_config['f_params']['a'],))
     train_loader_, test_loader, val_loader = data_gen.generate(
         N=dataset_config['num_batches'] * dataset_config['batch_size'], batch_size=dataset_config['batch_size'],
         splits=dataset_config['splits'])
@@ -233,7 +192,7 @@ if __name__ == '__main__':
     train_params_ = {'n_epochs': 1 if args.dryrun else 100, 'batch_size': dataset_config['batch_size'], 'lr': 1e-3,
                      't_span': dataset_config['t_span']}
 
-    ode_func_model_init = ODEFunc(device=torch_configs['device'], tensor_dtype=torch_configs['TENSOR_DTYPE'])
+    ode_func_model_init = ODEFuncNN3Layer(device=torch_configs['device'], tensor_dtype=torch_configs['TENSOR_DTYPE'])
     epochs_print_freq = max(int(train_params_['n_epochs'] / 10), 1)
     loss_fn = torch.nn.SmoothL1Loss()
     torch_solver_ = TorchRK45(device=torch_configs['device'], tensor_dtype=torch_configs['TENSOR_DTYPE'])
@@ -252,11 +211,12 @@ if __name__ == '__main__':
     end_time = datetime.datetime.now()
     # log results
     t_delta_fmt = format_timedelta(time_delta=end_time - start_time)
-    torch_trainer.save_model(model_path=os.path.join(args.model_dir, f'{args.model_prefix}_{experiment_timestamp}.model'))
+    torch_trainer.save_model(
+        model_path=os.path.join(args.model_dir, f'{args.model_prefix}_{experiment_timestamp}.model'))
     log_train_experiment(
         experiment_log_filepath=os.path.join(args.log_dir, f"train_experiment_{experiment_timestamp}.json"),
         run_type="train", solver=torch_solver_, f_true_dynamic=dataset_config['f_true_dynamics'],
-        ode_func=ODEFunc, training_time_fmt=t_delta_fmt, torch_config=torch_configs, data_config=dataset_config,
+        ode_func=ODEFuncNN3Layer, training_time_fmt=t_delta_fmt, torch_config=torch_configs, data_config=dataset_config,
         train_params=train_params_, train_loss=train_loss, nfe=torch_trainer.get_nfe(),
         total_solve_calls=torch_trainer.get_total_solve_call_count())
